Move create_report to instance method and run it at end of generate
