+++ src/gretel_trainer/relational/report/report.py
@@ -10,21 +10,8 @@ import pandas as pd
 from jinja2 import Environment, FileSystemLoader
 
 from gretel_trainer.relational.core import RelationalData, TableEvaluation
-from gretel_trainer.relational.multi_table import MultiTable
 
 _TEMPLATE_DIR = str(Path(__file__).parent)
-
-
-def create_report(multitable: MultiTable) -> None:
-    presenter = ReportPresenter(
-        rel_data=multitable.relational_data,
-        evaluations=multitable.evaluations,
-        now=datetime.datetime.utcnow(),
-    )
-    output_path = multitable._working_dir / "relational_report.html"
-    with open(output_path, "w") as report:
-        html_content = ReportRenderer().render(presenter)
-        report.write(html_content)
 
 
 class ReportRenderer:
